Fix failing tests
Changed the argument handling from what appeared to be command-line
based (where the 0th argument would be the command run) to a
framework-like call with only a list of arguments. There may be a
reason for this to be as it was that isn't apparent, but the default
handler should bypass a command line argument if used in that manner.